backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
PyPDF2
langchain
openai
duckduckgo-search
python-dotenv